mock backend info
