+++ manager/service.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"log/slog"
 	"net"
-	"os"
-	"os/exec"
 	"regexp"
 	"strconv"
 	"sync"
@@ -174,22 +172,6 @@ func (ms *managerService) Stop(ctx context.Context, computationID string) error 
 	}
 	delete(ms.vms, computationID)
 	return nil
-}
-
-func (ms *managerService) FetchBackendInfo() ([]byte, error) {
-	cmd := exec.Command("sudo", fmt.Sprintf("%s/backend_info", ms.backendMeasurementBinaryPath), "--policy", "1966081")
-
-	_, err := cmd.Output()
-	if err != nil {
-		return nil, err
-	}
-
-	f, err := os.ReadFile("./backend_info.json")
-	if err != nil {
-		return nil, err
-	}
-
-	return f, nil
 }
 
 func getFreePort(minPort, maxPort int) (int, error) {
